+++ src/bunit.web.query/Labels/Strategies/LabelTextUsingAriaLabelledByStrategy.cs
@@ -5,7 +5,7 @@ namespace Bunit.Labels.Strategies;
 
 internal sealed class LabelTextUsingAriaLabelledByStrategy : ILabelTextQueryStrategy
 {
-	public IElement? FindElement(IRenderedComponent<IComponent> renderedComponent, string labelText, ByLabelTextOptions options)
+	public IEnumerable<IElement> FindElements(IRenderedComponent<IComponent> renderedComponent, string labelText, ByLabelTextOptions options)
 	{
 		var elementsWithAriaLabelledBy = renderedComponent.Nodes.TryQuerySelectorAll("[aria-labelledby]");
 
@@ -13,9 +13,7 @@ internal sealed class LabelTextUsingAriaLabelledByStrategy : ILabelTextQueryStra
 		{
 			var labelElement = renderedComponent.Nodes.TryQuerySelector($"#{element.GetAttribute("aria-labelledby")}");
 			if (labelElement is not null && labelElement.GetInnerText().Equals(labelText, options.ComparisonType))
-				return element.WrapUsing(new ByLabelTextElementFactory(renderedComponent, labelText, options));
+				yield return element.WrapUsing(new ByLabelTextElementFactory(renderedComponent, labelText, options));
 		}
-
-		return null;
 	}
 }