+++ src/bunit.web.query/Labels/Strategies/LabelTextUsingWrappedElementStrategy.cs
@@ -5,15 +5,18 @@ namespace Bunit.Labels.Strategies;
 
 internal sealed class LabelTextUsingWrappedElementStrategy : ILabelTextQueryStrategy
 {
-	public IElement? FindElement(IRenderedComponent<IComponent> renderedComponent, string labelText, ByLabelTextOptions options)
+	public IEnumerable<IElement> FindElements(IRenderedComponent<IComponent> renderedComponent, string labelText, ByLabelTextOptions options)
 	{
-		var matchingLabel = renderedComponent.Nodes.TryQuerySelectorAll("label")
-			.SingleOrDefault(l => l.GetInnerText().Trim().StartsWith(labelText, options.ComparisonType));
+		var matchingLabels = renderedComponent.Nodes.TryQuerySelectorAll("label")
+			.Where(l => l.GetInnerText().Trim().StartsWith(labelText, options.ComparisonType));
 
-		var matchingElement = matchingLabel?
-			.Children
-			.SingleOrDefault(n => n.IsHtmlElementThatCanHaveALabel());
-
-		return matchingElement?.WrapUsing(new ByLabelTextElementFactory(renderedComponent, labelText, options));
+		foreach (var matchingLabel in matchingLabels)
+		{
+			var matchingElements = matchingLabel.Children.Where(n => n.IsHtmlElementThatCanHaveALabel());
+			foreach (var matchingElement in matchingElements)
+			{
+				yield return matchingElement.WrapUsing(new ByLabelTextElementFactory(renderedComponent, labelText, options));
+			}
+		}
 	}
 }